+++ src/client/Microsoft.Identity.Client/Properties/InternalsVisibleTo.cs
@@ -13,7 +13,7 @@ using Microsoft.Identity.Client;
 [assembly: InternalsVisibleTo("Microsoft.Identity.Test.Integration.NetCore" + KeyTokens.MSAL)]
 [assembly: InternalsVisibleTo("Microsoft.Identity.Test.Integration.NetFx" + KeyTokens.MSAL)]
 [assembly: InternalsVisibleTo("Microsoft.Identity.Test.Performance" + KeyTokens.MSAL)]
-[assembly: InternalsVisibleTo("Microsoft.Identity.Test.E2E.NetCore" + KeyTokens.MSAL)]
+[assembly: InternalsVisibleTo("Microsoft.Identity.Test.E2E.MSI" + KeyTokens.MSAL)]
 
 [assembly: InternalsVisibleTo("CommonCache.Test.Common" + KeyTokens.MSAL)]
 [assembly: InternalsVisibleTo("CommonCache.Test.Unit" + KeyTokens.MSAL)]
